first pass at reorging the files
