+++ libs/constants.py
@@ -1,0 +1,1 @@
+NULL_VALUE = "<Null>"
